feat: question 엔드포인트 생성 (모든 질문을 조회하는 프로시저를 수행하는 컨트롤러)
- 프로시저를 호출하는 api(엔드포인트)와 프로시저를 호출하는 메소드 구현이 필요.

- 프로시저를 호출하는 api는 프로시저를 호출하는 메소드를 호출하며 내부에서 로직을 작성해도 상관없음

+++ src/main/java/com/example/androidserver/Controller/UserController.java
@@ -31,7 +31,7 @@ public class UserController {
         @Override
         public User mapRow(ResultSet rs, int rowNum) throws SQLException {
             User user = new User();
-            user.setUid(rs.getLong("uid"));
+            user.setUid(rs.getInt("uid"));
             user.setEmail(rs.getString("email"));
             user.setInterest(rs.getString("interest"));
             return user;
